domain: moved conversion code to domain crate.
Issue: #74
