+++ block-producer/src/store/mod.rs
@@ -5,10 +5,8 @@ use std::{
 
 use async_trait::async_trait;
 use miden_node_proto::{
-    account,
-    conversion::convert,
-    digest,
-    domain::BlockInputs,
+    account, convert, digest,
+    domain::block_inputs::BlockInputs,
     requests::{ApplyBlockRequest, GetBlockInputsRequest, GetTransactionInputsRequest},
     store::api_client as store_client,
 };
