+++ proto/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod conversion;
 pub mod domain;
 pub mod errors;
 mod formatting;
@@ -9,6 +8,7 @@ mod generated;
 
 // RE-EXPORTS
 // ------------------------------------------------------------------------------------------------
+pub use domain::{convert, nullifier_value_to_blocknum};
 pub use generated::{
     account, block_header, block_producer, digest, merkle, mmr, note, requests, responses, rpc,
     smt, store,
